requirements.txt
# для работы с PostgreSQL
psycopg2-binary==2.9.11

# для красивого меню
rich==14.3.2